Split out interface. Added admin and confirmed constraints.
